requirements.txt
flask>=3.0
pyyaml>=6.0
requests>=2.31
